# v2/openapi/itips-v2.yaml
openapi: 3.1.0
info:
  title: ITIPS V2 — Jetson Edge API (Dahua-native)
  version: 2.1.0
  description: |
    Two HTTP surfaces run on every Jetson:

    * **Public 5050** — operator dashboard, MJPEG feeds, SSE alert stream,
      worker-enrollment / plate-list management, Test Console.
    * **Inbound 8443** — backend → Jetson commands (B1–B5). Bearer auth.

    All AI runs on the Dahua cameras. The Jetson subscribes to camera
    events, packages signed evidence, and proxies operator clicks onto
    `faceRecognitionServer.cgi`, `recordUpdater.cgi`, `coaxialControlIO.cgi`,
    and `ptz.cgi`.
servers:
  - url: http://localhost:5050
    description: Public surface
  - url: http://localhost:8443
    description: Inbound surface (bearer token required)
...
security: []
paths:
...
  /video_feed/{camera_id}:
    get:
      tags: [Snapshots]
      summary: MJPEG feed of the latest annotated snapshot
      description: |
        The Jetson no longer decodes RTSP continuously. The feed shows the
        most-recent snapshot fetched in response to a camera event (Dahua
        `/cgi-bin/snapshot.cgi`).
      parameters:
        - { name: camera_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: multipart/x-mixed-replace MJPEG stream
          content:
            multipart/x-mixed-replace: {}
...
  /local/api/v1/personnel/sync:
    post:
      tags: [Inbound]
      summary: B1 — sync a worker face into every camera
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/B1Payload" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/B1Result" }
  /local/api/v1/config:
    post:
      tags: [Inbound]
      summary: B2 — hot config push (acknowledged, no-op at Jetson level)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  applied: { type: boolean }
                  version: { type: string }
  /local/api/v1/maintenance/window:
    post:
      tags: [Inbound]
      summary: B3 — arm/disarm a maintenance window
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/B3Payload" }
      responses:
        "200": { description: OK }
  /local/api/v1/commands:
    post:
      tags: [Inbound]
      summary: B4 — PTZ override, deterrence fire/standdown, stream request
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/B4Payload" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  command_accepted: { type: boolean }
                  reason: { type: string }
  /local/api/v1/firmware/update:
    post:
      tags: [Inbound]
      summary: B5 — schedule firmware install
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: OK
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    B1Payload:
      type: object
      required: [action, person_id]
      properties:
        action: { type: string, enum: [add, update, deactivate] }
        person_id: { type: string }
        full_name: { type: string }
        image_b64:
          type: string
          description: |
            Base64-encoded JPEG. Required for `add` / `update`. The Jetson
            forwards the bytes to each camera's
            `faceRecognitionServer.cgi?action=addPerson`.
        sex: { type: string, enum: [Male, Female, Unknown] }
        birthday: { type: string, example: "1990-01-15" }
    B1Result:
      type: object
      properties:
        synced: { type: boolean }
        person_id: { type: string }
        cameras:
          type: object
          additionalProperties: { type: string }
        cameras_cleared:
          type: array
          items: { type: integer }
        failures:
          type: array
          items: { type: string }
    B3Payload:
      type: object
      required: [action, window_id, person_id]
      properties:
        action: { type: string, enum: [arm, disarm] }
        window_id: { type: string }
        person_id: { type: string }
        start_utc: { type: string }
        end_utc:   { type: string }
    B4Payload:
      type: object
      required: [command_type, parameters]
      properties:
        command_type:
          type: string
          enum: [ptz_override, deterrence_fire, deterrence_standdown, request_stream]
        parameters:
          type: object
          properties:
            camera_id: { type: integer }
            preset_id: { type: integer }
            pan: { type: number }
            tilt: { type: number }
            zoom: { type: number }
            light: { type: boolean }
            speaker: { type: boolean }
            bbox:
              type: array
              minItems: 4
              maxItems: 4
              items: { type: number }
            frame_width: { type: integer }
            frame_height: { type: integer }